refactored bloom filter optimizer out from array join optimizer
